fix: перенос из filmstorage
Добавлен likestorage и перенесены методы из filmstorage. Также относиться к mpa,genre storage

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -38,7 +38,7 @@ public class Film {
 
     @NotNull
     private final Mpa mpa;
-    private final Set<Long> likesFilm = new HashSet<>();
+    private final Set<Integer> likesFilm = new HashSet<>();
     private final Set<Genre> genres = new LinkedHashSet<>();
 
 }
